Added a category enum which we can define the specific categories which the user will be able to select.
Added the ReportTimes enum so we can tell the server for how long back we want to see the reports.

Removed the DataConnector class because we do not access the database directly from the client.

Added the report class which contains the methods to get and set the latitude, longitude, and category. It also has the method "addReport()" which will send the current report up to the database.